Added signal-to-noise ratio analyser (not working yet!)

+++ purview-analysers/snr-analyser/src/main/scala/org/purview/snranalyser/AnalyserImplementation.scala
@@ -14,11 +14,11 @@ class AnalyserImplementation extends HeatMapImageAnalyser{
 
   val grayscale: Computation[Matrix[Float]] = for (in <- input) yield (in.map(color => (color.r * 11f + color.g * 16f + color.b * 5f) / 32f))
 
-  def mean(matrix: Matrix[Float]): Float = matrix.sum / (matrix.width * matrix.height)
+  def mean(matrix: Matrix[Float]): Float = matrix.sum.toFloat / (matrix.width * matrix.height).toFloat
 
-  val meanGSC = for(gsc <- grayscale) yield (mean(gsc))
+  val meanGSC = for(gsc <- grayscale) yield (mean(gsc).toFloat)
 
-  val deviation = for(gsc <- grayscale; m <- meanGSC) yield (gsc map (_ - m))
+  val deviation = for(gsc <- grayscale; m <- meanGSC) yield (gsc map (_.toFloat - m.toFloat))
 
   val snr = for(m <- meanGSC; dev <- deviation) yield (dev map (x => m.toFloat / x.toFloat))
